feat(server): 增加管理员账户保护功能
- 新增 AdminConstant 类，定义管理员 ID常量
- 新增 AdminException 异常类，用于处理管理员相关异常
- 修改 MemberService接口，增加根据 ID 获取用户的方法
- 实现 MemberServiceImpl 中的管理员保护逻辑：
  -禁止删除管理员账户 - 禁止修改管理员账户
  - 新增用户时，检查用户名是否已存在
- 更新相关控制器和 Mapper 接口以支持新功能

+++ budget-pojo/src/main/java/com/budget/vo/AccountVO.java
@@ -17,7 +17,7 @@ public class AccountVO {
 
     private String  name;
 
-    private int typeId;
+    private Long typeId;
 
     private Long categoryId;
